Fixed building using MSVC not working
- Fixed some warnings

+++ Fusion/Source/Fusion/Renderer/Mesh.cpp
@@ -7,7 +7,7 @@ namespace Fusion {
 		: m_Vertices(InVertices), m_Indices(InIndices)
 	{
 		VertexBufferInfo CreateInfo;
-		CreateInfo.BufferSize = m_Vertices.size() * sizeof(Vertex);
+		CreateInfo.BufferSize = uint32_t(m_Vertices.size()) * sizeof(Vertex);
 		CreateInfo.Data = static_cast<void*>(m_Vertices.data());
 		CreateInfo.Layout = {
 			{ "POSITION", ShaderDataType::Float3, 0 },
@@ -16,7 +16,7 @@ namespace Fusion {
 		};
 		CreateInfo.Usage = EBufferUsage::Immutable;
 		m_VertexBuffer = VertexBuffer::Create(CreateInfo);
-		m_IndexBuffer = IndexBuffer::Create(m_Indices.size() * 3 * sizeof(uint32_t), m_Indices.data());
+		m_IndexBuffer = IndexBuffer::Create(uint32_t(m_Indices.size()) * 3 * sizeof(uint32_t), m_Indices.data());
 	}
 
 	Mesh::~Mesh()
